+++ src/lib/loader.ts
@@ -1,4 +1,4 @@
-import { anonymize, identity, parseBool } from './helpers';
+import { parseBool } from './helpers';
 import { createConfigStorage } from './storage';
 
 export interface ConfigLoaderOptions {
@@ -13,54 +13,50 @@ const defaultOpts: ConfigLoaderOptions = {
     envMode: 'default',
 };
 
-export const createAtomLoaderFactory = (storage: Record<any, any>, anonymize: (x: any) => any = identity) => {
-    const load = <T, N extends boolean, R = N extends false ? T : T | null>(
-        transform: (x: any) => T,
-        anonymize: (x: any) => any,
-        hidden: boolean,
-        nullable: boolean
-    ) => (key: string): R => {
+export interface LoadedValue<T, N extends boolean, R = N extends false ? T : T | null> {
+    rawValue: any;
+    value: R;
+    hidden: boolean;
+    nullable: boolean;
+    __CONFIGURU_LEAF: true;
+}
+
+export const createAtomLoaderFactory = (storage: Record<any, any>) => {
+    const load = <T, N extends boolean>(transform: (x: any) => T, hidden: boolean, nullable: boolean) => (
+        key: string
+    ): LoadedValue<T, N> => {
         const value = storage[key];
-        if (value === undefined || value === null) {
-            if (nullable) {
-                return value;
-            }
+        const missing = value === undefined || value === null;
+        if (!nullable && missing) {
             throw new Error(`Missing required value ${key}`);
         }
-        if (hidden) {
-            return anonymize(value);
-        }
-        return (transform(value) as any) as R;
+        return {
+            hidden,
+            nullable,
+            rawValue: value,
+            value: missing ? null : (transform(value) as any),
+            __CONFIGURU_LEAF: true,
+        };
     };
     return <T>(transform: (x: any) => T) =>
-        Object.assign(load<T, false>(transform, anonymize, false, false), {
-            hidden: Object.assign(load<T, false>(transform, anonymize, true, false), {
-                nullable: load<T, true>(transform, anonymize, true, true),
+        Object.assign(load<T, false>(transform, false, false), {
+            hidden: Object.assign(load<T, false>(transform, true, false), {
+                nullable: load<T, true>(transform, true, true),
             }),
-            nullable: Object.assign(load<T, true>(transform, anonymize, false, true), {
-                hidden: load<T, true>(transform, anonymize, true, true),
+            nullable: Object.assign(load<T, true>(transform, false, true), {
+                hidden: load<T, true>(transform, true, true),
             }),
         });
 };
 
-export const createLoaders = (opts: ConfigLoaderOptions = defaultOpts) => {
+export const createLoader = (opts: ConfigLoaderOptions = defaultOpts) => {
     const configStorage = createConfigStorage(opts);
     const atomLoader = createAtomLoaderFactory(configStorage);
-    const anonymousAtomLoader = createAtomLoaderFactory(configStorage, anonymize);
     return {
-        configLoader: {
-            number: atomLoader(Number),
-            string: atomLoader(String),
-            bool: atomLoader(parseBool),
-            json: atomLoader(JSON.parse),
-        },
-        anonymizedConfigLoader: {
-            number: anonymousAtomLoader(Number),
-            string: anonymousAtomLoader(String),
-            bool: anonymousAtomLoader(parseBool),
-            json: anonymousAtomLoader(JSON.parse),
-        },
+        number: atomLoader(Number),
+        string: atomLoader(String),
+        bool: atomLoader(parseBool),
+        json: atomLoader(JSON.parse),
+        custom: <T>(fn: (x: any) => T) => atomLoader(fn),
     };
 };
-
-export type ConfigLoader = ReturnType<typeof createLoaders>['configLoader'];