+++ src/test/atomLoader.test.ts
@@ -7,33 +7,79 @@ const config = {
     null: null,
 };
 
-const stringLoader = createAtomLoaderFactory(config, () => '***')(x => String(x));
+const stringLoader = createAtomLoaderFactory(config)(x => String(x));
 
 describe('atomLoader', () => {
     describe('string loader', () => {
         test('number', () => {
-            expect(stringLoader('number')).toBe('42');
+            expect(stringLoader('number')).toMatchInlineSnapshot(`
+                Object {
+                  "__CONFIGURU_LEAF": true,
+                  "hidden": false,
+                  "nullable": false,
+                  "rawValue": 42,
+                  "value": "42",
+                }
+            `);
         });
         test('nullable', () => {
             // empty throws on default
             expect(() => stringLoader('null')).toThrow(/missing/i);
             expect(() => stringLoader('undefined')).toThrow(/missing/i);
             // empty valid on nullable
-            expect(stringLoader.nullable('null')).toBe(null);
-            expect(stringLoader.nullable('undefined')).toBe(undefined);
+            expect(stringLoader.nullable('null')).toMatchInlineSnapshot(`
+                Object {
+                  "__CONFIGURU_LEAF": true,
+                  "hidden": false,
+                  "nullable": true,
+                  "rawValue": null,
+                  "value": null,
+                }
+            `);
+            expect(stringLoader.nullable('undefined')).toMatchInlineSnapshot(`
+                Object {
+                  "__CONFIGURU_LEAF": true,
+                  "hidden": false,
+                  "nullable": true,
+                  "rawValue": undefined,
+                  "value": null,
+                }
+            `);
         });
         test('nullable & hidden', () => {
             // empty throws on hidden
             expect(() => stringLoader.hidden('null')).toThrow(/missing/i);
             expect(() => stringLoader.hidden('undefined')).toThrow(/missing/i);
             // empty valid on nullable
-            expect(stringLoader.hidden.nullable('null')).toBe(null);
-            expect(stringLoader.hidden.nullable('undefined')).toBe(undefined);
+            expect(stringLoader.hidden.nullable('null')).toMatchInlineSnapshot(`
+                Object {
+                  "__CONFIGURU_LEAF": true,
+                  "hidden": true,
+                  "nullable": true,
+                  "rawValue": null,
+                  "value": null,
+                }
+            `);
+            expect(stringLoader.hidden.nullable('undefined')).toMatchInlineSnapshot(`
+                Object {
+                  "__CONFIGURU_LEAF": true,
+                  "hidden": true,
+                  "nullable": true,
+                  "rawValue": undefined,
+                  "value": null,
+                }
+            `);
         });
         test('hidden', () => {
-            expect(stringLoader.hidden('number')).toBe('***');
-            expect(stringLoader.hidden('string')).toBe('***');
-            expect(stringLoader.hidden('boolean')).toBe('***');
+            expect(stringLoader.hidden('string')).toMatchInlineSnapshot(`
+                Object {
+                  "__CONFIGURU_LEAF": true,
+                  "hidden": true,
+                  "nullable": false,
+                  "rawValue": "string",
+                  "value": "string",
+                }
+            `);
         });
     });
 });